DMP-2258 move 2 methods to stepdef_base
